+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.9.9
 
 - Add new spell for Holy Wands: Heal
-- Reduce range of Flash Heal 
+- Reduce range of Flash Heal
+- Rework Divine Protection spell
 - Fix warnings recipes with absent materials
 - Update JSON API usage